LAB-1930: Read foreground pgrp from PTY master on macOS
macOS rejects TIOCGPGRP with ENOTTY on a freshly opened slave tty
path (it is already the controlling terminal of the shell's session),
so the darwin foregroundProcessGroup() always errored. The err!=nil
fallback then forced the foreground pgrp to the shell pgrp, marking
every pane idle -- so capture reported exited=true and
current_command="bash" even while htop/nvim/claude ran.

Read TIOCGPGRP from the PTY master fd instead (as linux already did),
and consolidate both platforms into one implementation in
pane_signal_unix.go. The darwin-only slave-reopen path and its
TIOCPTYGNAME/ttyPath helpers are removed.

Regression covered by the pre-existing TestAgentStatusTracksBusyAndIdle,
which failed before this change and passes after.
